requirements.txt
autoenv
Flask
pyjwt
pandas
numpy
requests